update calculation of relative IQ loss
Following a suggestion from Heather, we simplified our calculation of
relative IQ loss. Rather than discounting a lifetime figure in dollars
of foregone earnings in the US and then converting this to a percentage
by dividing by US GDP/capita, yielding a 1.3% figure, we proceeded
directly to the percentage figure from which the lifetime figure in US
dollars was calcualated. This percentage was 2.6% and comes from a 2019
EPA documented commented in our code. This is not only simpler, but more
accurate. It doubles our overall result.