use new BLL to IQ loss function
On the advice of the referees we've switched from the piecewise linear
BLL to IQ loss function of Lanphear et al (2005) to the log-linear
function of Crump et al (2013). The results are qualitatively similar,
but the Crump function involves a higher estimated loss from the same
BLL, so the total estimated costs have increased by something like a
factor of 2.

This commit replaces the old function with the new one and re-runs the
whole pipeline. It also makes a small change to the 5th part of the
pipeline in which we compute numerical summaries of our plots. These
have been split off into a separate 6th step to make it clearer which
script does what.