robustness check: scaling up BLLs
This robustness check proceeds by scaling up the raw average BLL data
from the GBD by a fixed multiplier, e.g. 1.05 for a 5% increase. We then
impute the new values of the fraction of people with BLLs >= 5 and 10
micrograms per deciliters using a pair of LOESS regressions of average
BLL and the fractions above these threshold from the real GBD data.

After constructing this "scaled-up" version of the BLL data, we proceed
with the same pipepline as before. Finally we store the scaled-up
results and then compare them to the original results to see how much of
a difference it makes.